Move entities serializers
