+++ app/experimenter/normandy/serializers.py
@@ -7,7 +7,7 @@ from experimenter.experiments.models import (
     VariantPreferences,
     RolloutPreference,
 )
-from experimenter.experiments.serializers.entities import PrefTypeField
+from experimenter.experiments.api.v1.serializers import PrefTypeField
 from experimenter.experiments.constants import ExperimentConstants
